requirements.txt
python-telegram-bot
apscheduler
python-dotenv
pytest
